lifecycle for run app

+++ src/daemon/rest/protoc/ProtobufHelper.cpp
@@ -66,7 +66,7 @@ int ProtobufHelper::readMsgHeader(const ACE_SSL_SOCK_Stream &socket, ssize_t &re
 	return bodySize;
 }
 
-const std::tuple<std::shared_ptr<char>, int> ProtobufHelper::readBytes(const ACE_SSL_SOCK_Stream &socket, int bodySize, ssize_t &recvReturn)
+const std::tuple<std::shared_ptr<char>, int> ProtobufHelper::readBytes(const ACE_SSL_SOCK_Stream &socket, size_t bodySize, ssize_t &recvReturn)
 {
 	const static char fname[] = "ProtobufHelper::readBytes() ";
